Integrate test
